requirements.txt
# Core Dependencies
langgraph==0.2.59
anthropic==0.57.1
python-dotenv==1.0.1

# Development Dependencies
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0
black==24.10.0
pylint==3.3.3
mypy==1.14.0
types-requests==2.32.0.20241016

# Additional utilities
tenacity==9.0.0  # For retry logic
beautifulsoup4==4.12.3  # For HTML parsing
requests==2.32.3  # For HTTP requests

# Observability
langsmith==0.1.147  # For LangGraph tracing and monitoring